+++ tool/grind.dart
@@ -28,6 +28,8 @@ void main(List<String> args) {
   pkg.homebrewFormula.value = "sass.rb";
   pkg.jsRequires.value = {"chokidar": "chokidar", "readline": "readline"};
   pkg.jsModuleMainLibrary.value = "lib/src/node.dart";
+  pkg.npmAdditionalFiles.fn = () =>
+      ({"postinstall.js": File("package/postinstall.js").readAsStringSync()});
   pkg.npmPackageJson.fn = () =>
       json.decode(File("package/package.json").readAsStringSync())
           as Map<String, dynamic>;